Rust-ify fmt::Display for Representation
One neat thing about Rust is that `?` convention does exactly what you were doing earlier with the `if let`s, at least in any function with a `Result` return signature. If there are any errors, return that. Otherwise proceed as normal.

I ran `cargo run` on my machine to verify that it works, but let me know if it works for you.